Moved files into more structured system.
