+++ JetLagged/app/src/main/java/com/example/jetlagged/data/JetLaggedHomeScreenViewModel.kt
@@ -1,7 +1,5 @@
-package com.example.jetlagged
+package com.example.jetlagged.data
 
-import androidx.compose.runtime.MutableState
-import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.ViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
