+++ JetLagged/app/src/main/java/com/example/jetlagged/sleep/SleepBar.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.example.jetlagged
+package com.example.jetlagged.sleep
 
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.ExperimentalAnimationApi
@@ -77,6 +77,7 @@ import androidx.compose.ui.text.rememberTextMeasurer
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.util.lerp
+import com.example.jetlagged.data.sleepData
 import com.example.jetlagged.ui.theme.LegendHeadingStyle
 
 @OptIn(ExperimentalAnimationApi::class)
